new way of adding headers
